requirements.txt
discord.py==1.2.3
pytest==5.0.0
cryptography==2.7
markovify==0.7.1
numpy==1.16.4
nltk==3.3
scikit-learn==0.19.2
scipy==1.1.0